#333 Extract a shared 4bpp test harness
